refactor: corrected typo in FindAllSystematicStudiesService
