+++ review/src/main/kotlin/br/all/application/review/find/presenter/FindAllSystematicStudyPresenter.kt
@@ -1,6 +1,6 @@
 package br.all.application.review.find.presenter
 
-import br.all.application.review.find.services.FindAllSystematicStudyService.ResponseModel
+import br.all.application.review.find.services.FindAllSystematicStudiesService.ResponseModel
 import br.all.application.shared.presenter.GenericPresenter
 
 interface FindAllSystematicStudyPresenter: GenericPresenter<ResponseModel>
