+++ review/src/test/kotlin/br/all/application/review/util/TestDataFactory.kt
@@ -9,7 +9,7 @@ import io.github.serpro69.kfaker.Faker
 import java.util.*
 import br.all.application.review.create.CreateSystematicStudyService.RequestModel as CreateRequestModel
 import br.all.application.review.create.CreateSystematicStudyService.ResponseModel as CreateResponseModel
-import br.all.application.review.find.services.FindAllSystematicStudyService.ResponseModel as FindAllResponseModel
+import br.all.application.review.find.services.FindAllSystematicStudiesService.ResponseModel as FindAllResponseModel
 import br.all.application.review.find.services.FindOneSystematicStudyService.ResponseModel as FindOneResponseModel
 
 class TestDataFactory {
