+++ web/src/main/kotlin/br/all/review/presenter/RestfulFindAllSystematicStudiesPresenter.kt
@@ -1,7 +1,7 @@
 package br.all.review.presenter
 
 import br.all.application.review.find.presenter.FindAllSystematicStudyPresenter
-import br.all.application.review.find.services.FindAllSystematicStudyService
+import br.all.application.review.find.services.FindAllSystematicStudiesService
 import br.all.application.review.repository.SystematicStudyDto
 import br.all.review.controller.SystematicStudyController
 import br.all.shared.error.createErrorResponseFrom
@@ -14,7 +14,7 @@ import java.util.*
 class RestfulFindAllSystematicStudiesPresenter: FindAllSystematicStudyPresenter {
     var responseEntity: ResponseEntity<*>? = null
 
-    override fun prepareSuccessView(response: FindAllSystematicStudyService.ResponseModel) {
+    override fun prepareSuccessView(response: FindAllSystematicStudiesService.ResponseModel) {
         val restfulResponse = ViewModel(
             response.researcherId,
             response.systematicStudies.size,
